Add create_user layout, activity, and API call
This commit adds a new layout file called create_user.xml, along with a corresponding activity called CreateUserActivity that handles the creation of a new user on the server. The activity sends a POST request to https://dummyjson.com/users/add to add the new user. Additionally, the MainActivity has been updated to open CreateUserActivity when the corresponding navigation menu item is clicked.