Add ProductViewActivity with image viewing and management features
This commit adds a new activity called ProductViewActivity, which allows the user to view a selected product with images that can be swiped left or right. The activity also implements GestureDetector.OnGestureListener for additional touch functionality, including the ability to long-hold an image to save it to local storage. The app requests the necessary permissions based on the device's API level.

In addition to viewing the product, the user can also manage it using two imageViews that have set onClick listeners. Clicking on the first imageView will delete the product from the server using JsonDeleteProduct class, which sends the request using the okhttp3 library and extends AsyncTask. Clicking on the second imageView opens a new activity called UpdateProductActivity, which uses the update_product_layout and allows the user to update the product on the server using JsonUpdateProduct class, which also uses the okhttp3 library and extends AsyncTask. The activity also displays product title, description, rating, stock, and price."