Add UserViewActivity and user view layout file
This commit adds a new activity called UserViewActivity, along with the corresponding activity_user_view.xml layout file. When a user is clicked, the activity opens and receives an extra User object to display the user's data.

In addition to displaying user data, the commit also adds a new drawable for the userRemove imageView and sets up an onClick listener to delete the user from the server. Furthermore, a new TextView has been added to open a new activity to update user data on the server.