Add UpdateUserActivity with layout file
This commit adds a new class called UpdateUserActivity, which updates user data on the server. It also includes a new layout file called update_user_layout for this activity.

Additionally, the UserViewActivity class has been updated to open UpdateUserActivity when the user clicks on the TextView. The JsonUpdateUser class is called when the TextView is clicked, and updates the user data on the server